requirements.txt
gradio
openai
mem0ai
python-dotenv